+++ laktory/models/resources/databricks/job.py
@@ -374,10 +374,16 @@ class JobTaskSqlTaskFile(BaseModel):
     Attributes
     ----------
     path:
-        SQL filepath
+        If source is `GIT`: Relative path to the file in the repository specified
+        in the git_source block with SQL commands to execute. If source is
+        `WORKSPACE`: Absolute path to the file in the workspace with SQL
+        commands to execute.
+    source:
+        The source of the project. Possible values are `WORKSPACE` and `GIT`.
     """
 
     path: str = None
+    source: Literal["WORKSPACE", "GIT"] = None
 
 
 class JobTaskSQLTask(BaseModel):
@@ -404,8 +410,8 @@ class JobTaskSQLTask(BaseModel):
     dashboard: JobTaskSqlTaskDashboard = None
     file: JobTaskSqlTaskFile = None
     parameters: dict[str, Any] = None
-    query: JobTaskSqlTaskQuery
-    warehouse_id: str = None
+    query: JobTaskSqlTaskQuery = None
+    warehouse_id: str
 
 
 class JobTask(BaseModel):
@@ -625,6 +631,8 @@ class Job(BaseModel, PulumiResource, TerraformResource):
         If `True`, the Databricks provider will stop and start the job as needed to ensure that the active run for the
         job reflects the deployed configuration. For continuous jobs, the provider respects the pause_status by
         stopping the current active run. This flag cannot be set for non-continuous jobs.
+    description:
+        An optional description for the job. The maximum length is 1024 characters in UTF-8 encoding.
     email_notifications:
         An optional set of email addresses notified when runs of this job begins, completes or fails. The default
         behavior is to not send any emails. This field is a block and is documented below.
@@ -717,6 +725,7 @@ class Job(BaseModel, PulumiResource, TerraformResource):
     clusters: list[JobCluster] = []
     continuous: JobContinuous = None
     control_run_state: bool = None
+    description: str = None
     email_notifications: JobEmailNotifications = None
     format: str = None
     health: JobHealth = None